+++ src/TBD_macros.jl
@@ -158,7 +158,7 @@ end
 function process_mutate_expression(expr, sq, select_expressions, cte_name)
     if isa(expr, Expr) && expr.head == :(=) && isa(expr.args[1], Symbol)
         col_name = string(expr.args[1])
-        if current_sql_mode[] == :snowflake
+        if current_sql_mode[] == snowflake()
             col_name = uppercase(col_name)
         end
         col_expr = expr_to_sql(expr.args[2], sq)  # Convert to SQL expression
@@ -373,7 +373,7 @@ function process_summary_expression(expr, sq, summary_str)
         summary_operation = string(summary_operation)
         summary_column = expr_to_sql(expr.args[1], sq, from_summarize = true)
         summary_column = string(summary_column)
-        if current_sql_mode[] == :snowflake
+        if current_sql_mode[] == snowflake()
             summary_column = uppercase(summary_column)
         end
         push!(sq.metadata, Dict("name" => summary_column, "type" => "UNKNOWN", "current_selxn" => 1, "table_name" => sq.from))
@@ -653,49 +653,53 @@ macro show_query(sqlquery)
     end
 end
 
-macro collect(sqlquery)
-    return quote
-        # Extract the database connection from the SQLQuery object
-        db = $(esc(sqlquery)).db
-        sq = $(esc(sqlquery))
-        # Finalize the query to get the SQL string
-        final_query = finalize_query($(esc(sqlquery)))
-        df_result = DataFrame()
-        # Determine the type of db and execute the query accordingly
-        if db isa DatabricksConnection
-            df_result = execute_databricks(db, final_query)
-        elseif db isa SQLite.DB || db isa LibPQ.Connection || db isa DuckDB.DB || db isa MySQL.Connection || db isa ODBC.Connection
-            result = DBInterface.execute(db, final_query)
-            df_result = DataFrame(result)
-        elseif current_sql_mode[] == :clickhouse
-            df_result = ClickHouse.select_df(db, final_query)
-            selected_columns_order = sq.metadata[sq.metadata.current_selxn .== 1, :name]
-            df_result = df_result[:, selected_columns_order]
-        elseif db isa GoogleSession{JSONCredentials}
-                df_result = collect_gbq(sq.db, final_query)
-        elseif current_sql_mode[] == :snowflake
-            df_result = execute_snowflake(db, final_query)
-        elseif current_sql_mode[] == :athena
-            exe_query = Athena.start_query_execution(final_query, sq.athena_params; aws_config = db)
-                status = "RUNNING"
-        while status in ["RUNNING", "QUEUED"]
-            sleep(1)  # Wait for 1 second before checking the status again to avoid flooding the API
-            query_status = Athena.get_query_execution(exe_query["QueryExecutionId"], sq.athena_params; aws_config = db)
-            status = query_status["QueryExecution"]["Status"]["State"]
-            if status == "FAILED"
-                error("Query failed: ", query_status["QueryExecution"]["Status"]["StateChangeReason"])
-            elseif status == "CANCELLED"
-                error("Query was cancelled.")
-            end
-        end
-        
-        # Fetch the results once the query completes
-        result = Athena.get_query_results(exe_query["QueryExecutionId"], sq.athena_params; aws_config = db)
-            df_result = collect_athena(result)
-        else
-            error("Unsupported database type: $(typeof(db))")
-        end
-        df_result
-    end
+
+
+function final_collect(sqlquery::SQLQuery, ::Type{<:duckdb})
+    final_query = finalize_query(sqlquery)
+    result = DBInterface.execute(sqlquery.db, final_query)
+    return DataFrame(result)
 end
 
+function final_collect(sqlquery::SQLQuery, ::Type{<:databricks})
+    final_query = finalize_query(sqlquery)
+    result = execute_databricks(sqlquery.db, final_query)
+    return DataFrame(result)
+end
+
+function final_collect(sqlquery::SQLQuery, ::Type{<:snowflake})
+    final_query = finalize_query(sqlquery)
+    result = execute_snowflake(sqlquery.db, final_query)
+    return DataFrame(result)
+end
+
+macro collect(sqlquery)
+    return quote
+        backend = current_sql_mode[]
+        if backend == duckdb()
+            final_collect($(esc(sqlquery)), duckdb)
+        elseif backend == clickhouse()
+            final_collect($(esc(sqlquery)), clickhouse)
+        elseif backend == sqlite()
+            final_collect($(esc(sqlquery)), sqlite)
+        elseif backend == mysql()
+            final_collect($(esc(sqlquery)), mysql)
+        elseif backend == mssql()
+            final_collect($(esc(sqlquery)), mssql)
+        elseif backend == postgres()
+            final_collect($(esc(sqlquery)), postgres)
+        elseif backend == athena()
+            final_collect($(esc(sqlquery)), athena)
+        elseif backend == snowflake()
+            final_collect($(esc(sqlquery)), snowflake)
+        elseif backend == gbq()
+            final_collect($(esc(sqlquery)), gbq)
+        elseif backend == oracle()
+            final_collect($(esc(sqlquery)), oracle)
+        elseif backend == databricks()
+            final_collect($(esc(sqlquery)), databricks)
+        else
+            throw(ArgumentError("Unsupported SQL mode: $backend"))
+        end
+    end
+end